pyproject.toml
[project]
name = "air-github-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "helix-py>=0.2.31",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
